Окончательно развязал одиночный файл и zip-архив
